move lsp mux protocol extensions to their own module
